requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
reportlab>=4.0.0
Pillow>=10.0.0
pandas>=2.0.0
gspread>=6.0.0
google-auth>=2.0.0
